Split tests for inspect
